Dialog refactoring

+++ CharacterInformator.cs
@@ -6,7 +6,7 @@ public class CharacterInformator : ScriptableObject
     [Tooltip("Name of unit")]
     [SerializeField]
     private string[] unitName;
-    public string Name { get { return unitName[PlayerManager.Language]; } }
+    public string Name { get { return unitName[(int)PlayerManager.Language]; } }
 
     [Tooltip("Sprite of unit")]
     [SerializeField]
